Refactor hugepage out of tt_SiliconDevice (#253)
Related to #142 
Moved hugepage related code to a separate file and to PCIDevice, to
further relieve tt_SiliconDevice.

No API changes are imposed with this change.

- Following have been moved to separate compilation unit, which are OS
related: get_num_hugepages, get_available_num_host_mem_channels,
find_hugepage_dir, open_hugepage_file
- These have been moved to PCIDevice since they are lower level concepts
in our ideal hierarchy: init_hugepage
- Added hugepage_mapping struct to hold hugepage intialization related
info.
- ioctl and communication with kernel driver can now be completelly
removed from tt_SiliconDevice
